requirements.txt
playwright
pytest
pytest-playwright
allure-pytest
pyyaml
PyQt6
requests
configparser